requirements.txt
Django==5.1.5
django-allauth==65.3.0
django-crispy-forms==2.3
crispy-bootstrap4==2025.6
python-dotenv==1.0.1
Pillow==11.0.0
django-axes==8.1.0
bleach==6.3.0
django-ratelimit==4.1.0
django-cors-headers==4.9.0
django-csp==4.0
gunicorn==21.2.0
psycopg2-binary==2.9.10
django-redis==5.4.0
whitenoise==6.6.0
sentry-sdk==1.40.0
requests==2.32.3
PyJWT==2.10.1
asgiref==3.11.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==44.0.0
idna==3.10
oauthlib==3.2.2
packaging==24.2
pycparser==2.21
reportlab==4.2.5
requests-oauthlib==2.0.0
setuptools==75.8.0
sqlparse==0.5.3
urllib3==2.3.0
webencodings==0.5.1
djangorestframework==3.15.2
django-filter==24.3